Fix zeroc-ice33 optional dependency path generation in SUPERENV
With `ome/alt/zeroc-ice33` the determine_path() method of superenv.rb was
failing when adding the Ice binaries to the path. There are two solutions:
either reenabling the userpath to SUPERENV or using `zeroc-ice33` instead.

I will be using the second solution for now. However, the formula will need to
be retested with OMERO + Ice 3.4 .